finished read and archive
Read and archive are working. Found a bug however; in the inbox, when theres multiple messages, if you click one, it simply expands and if you click another, the text in the first message is replaced by the text in the second message.

Having them expand and collapse when clicked could be a cool feature though, just need then to display properly and collapse when clicked while expanded.